+++ crates/sui-framework/packages/sui-framework/sources/coin.move
@@ -165,6 +165,8 @@ public fun put<T>(balance: &mut Balance<T>, coin: Coin<T>) {
 // === Address Balance <-> Coin utility functions ===
 
 /// Redeem a `Withdrawal<Balance<T>>` and create a `Coin<T>` from the withdrawn Balance<T>.
+/// For an object withdrawal, the transaction aborts if the withdrawal specifies more funds than
+/// are currently available to the object.
 public fun redeem_funds<T>(
     withdrawal: sui::funds_accumulator::Withdrawal<Balance<T>>,
     ctx: &mut TxContext,
