+++ crates/sui-swarm-config/src/network_config_builder.rs
@@ -768,7 +768,9 @@ mod test {
                 epoch.epoch_start_timestamp(),
                 input_objects,
                 sui_types::base_types::SystemObjectVersions::default(),
-                None,
+                // The genesis transaction cannot withdraw object funds, so there are never
+                // unsettled withdrawals for it to account for.
+                &sui_types::accumulator_root::EmptyUnsettledObjectFunds,
                 gas_data,
                 SuiGasStatus::new_unmetered(),
                 kind,
