+++ crates/sui-types/src/accumulator_root.rs
@@ -137,6 +137,18 @@ pub trait UnsettledObjectFundsRead {
     ) -> u128;
 }
 
+pub struct EmptyUnsettledObjectFunds;
+
+impl UnsettledObjectFundsRead for EmptyUnsettledObjectFunds {
+    fn get_unsettled_object_withdraw(
+        &self,
+        _account: &AccumulatorObjId,
+        _accumulator_version: SequenceNumber,
+    ) -> u128 {
+        0
+    }
+}
+
 impl AccumulatorValue {
     pub fn as_u128(&self) -> Option<u128> {
         match self {
