+++ sui-execution/src/v2.rs
@@ -74,7 +74,7 @@ impl executor::Executor for Executor {
         epoch_timestamp_ms: u64,
         input_objects: CheckedInputObjects,
         _system_object_versions: sui_types::base_types::SystemObjectVersions,
-        _unsettled_object_funds: Option<&dyn sui_types::accumulator_root::UnsettledObjectFundsRead>,
+        _unsettled_object_funds: &dyn sui_types::accumulator_root::UnsettledObjectFundsRead,
         gas: GasData,
         gas_status: SuiGasStatus,
         transaction_kind: TransactionKind,
@@ -196,7 +196,7 @@ impl executor::Executor for Executor {
         epoch_timestamp_ms: u64,
         input_objects: CheckedInputObjects,
         _system_object_versions: sui_types::base_types::SystemObjectVersions,
-        _unsettled_object_funds: Option<&dyn sui_types::accumulator_root::UnsettledObjectFundsRead>,
+        _unsettled_object_funds: &dyn sui_types::accumulator_root::UnsettledObjectFundsRead,
         gas: GasData,
         gas_status: SuiGasStatus,
         transaction_kind: TransactionKind,
